requirements.txt
anndata==0.8.0
click==8.1.3
click-log==0.4.0
# deepcell==0.12.0  # install manually, to avoid version conflicts
fcswrite==0.6.1
h5py==3.2.1  # cellprofiler 4.2.1 requires 3.2.1 (note: deepcell 0.12.0 requires tensorflow ~=2.8.0, which technically requires ~=3.6.0)
imageio==2.19.3
networkx==2.8.2
numpy==1.22.4
opencv-python-headless==4.5.5.64
pandas==1.4.2
pyyaml==6.0
readimc==0.6.1
scikit-image==0.19.2
scipy==1.8.1
tifffile==2022.5.4
xtiff==0.7.8
